Fix 1 & 2
1- For whatever reason I thought it made sense for the y value of a sprite to be determined by the sprite's height. But since pixels are in a single array, the amount of rows we have is based on how many times the width is met.
TLDR, both X and Y are determined by modules.width, height is irrelevant.

2- Blurry textures caused by new Texture2D() defaults. Set the filter to point as opposed to billinear.

GenerateMasterSprite now has a verifier that you can use to make sure each generated sprite matches its intended module. You can select an image that exists in the project, then select a module name in the dropdown, and two objects will appear in the scene that will show you the chosen icon and the generated icon. And there's a button for destroying the objects once you're done.