Remove NameToData()
A bit more complicated than having two arrays, I've replaced the NameToData method with an OrderedDictionary, which works something similar to a Dictionary. This allows us to call the module's data directly using the name while also allowing us to use indexes when necessary for sprite matching.

This requires a rewrite for the Sprite Generator as it was designed to read a single line of names, as opposed to a multiline of names.

The next version of the generator should hopefully no longer require parsing the iconicScript directly, but we make do with what we have now.