My brain hurts.
Move Start back to Awake. This change was unimportant.

Move ModuleList out of Start and move it into Delay. There is no point in delaying if ModuleList is the thing we're waiting to not be null

Move adding needies into delay as well, no point in activating them if we don't have the list yet.

Use newtonsoft.json instead of JsonUtility. There ought to be a reason why the devs decided to include newtonsoft.json with every mod despite JsonUtility existing in Unity and given the fact it just doesn't work in game for who knows whatever reason, I don't see the need to dial back here.

Does Resources.Load even work here? Maybe it does but it doesn't work for me and I assume it's because Resources.Load is only meant to be loaded from the game's own Resources folder as opposed to from within an AssetBundle. As a matter of fact, Unity's documentation even points out that AssetBundles are a replacement for Resources.Load. Anyway we can include the textasset directly in the gameobject so why are we trying to load it